debug notify.py
