+++ src/etc/server.conf
@@ -1,5 +1,5 @@
 # debug, release
-RunMode = "release"
+RunMode = "debug"
 
 # my cluster name
 ClusterName = "Default"
@@ -11,7 +11,7 @@ EngineDelay = 120
 # log write dir
 Dir = "logs"
 # log level: DEBUG INFO WARNING ERROR
-Level = "INFO"
+Level = "DEBUG"
 # stdout, stderr, file
 Output = "stdout"
 # # rotate by time
