+++ src/etc/template/mailbody.tpl
